refactor: change app folder structure
