+++ module_3/datasource/db.js
@@ -1,5 +1,5 @@
 import knex from 'knex';
-import { config } from '../config';
+import { config } from '../../config';
 
 export const db = knex({
     client: 'postgres',
